move recentlyViewed code to workspace-utils
